AIEducationAll/backend/requirements.txt
# Core FastAPI and Server
fastapi>=0.100.0,<0.112.0 # Pin to a known compatible range or use latest
uvicorn[standard]>=0.20.0,<0.30.0 # ASGI server with extras like websockets, http-tools

# Data Validation and Settings
pydantic>=1.10.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0 # For loading settings from env vars

# Database (PostgreSQL with async)
sqlalchemy[asyncio]>=2.0.0,<2.1.0 # ORM, including async support
asyncpg>=0.25.0,<0.30.0 # Async PostgreSQL driver
alembic>=1.8.0,<2.0.0 # For database migrations (recommended)

# Security and Authentication
passlib[bcrypt]>=1.7.4,<2.0.0 # Password hashing
python-jose[cryptography]>=3.3.0,<4.0.0 # JWT handling

# Environment Variables
python-dotenv>=0.20.0,<2.0.0

# Optional: Background Tasks (if using Celery)
# celery>=5.0,<6.0
# redis>=4.0,<5.0 # Or rabbitmq library (e.g., pika)

# Optional: Production WSGI/ASGI Server (instead of just uvicorn for dev)
gunicorn>=20.0,<22.0